+++ scrapyd_client/cli.py
@@ -1,18 +1,67 @@
 import sys
 from argparse import ArgumentParser
+from textwrap import indent
 from traceback import print_exc
 
 import requests
 
-from scrapyd_client import commands
-from scrapyd_client.utils import (
-    DEFAULT_TARGET_URL,
-    ErrorResponse,
-    MalformedResponse,
-    get_config,
-)
+import scrapyd_client.deploy
+from scrapyd_client import lib
+from scrapyd_client.exceptions import ErrorResponse, MalformedResponse
+from scrapyd_client.utils import DEFAULT_TARGET_URL, get_config
 
+INDENT_PREFIX = "  "
 ISSUE_TRACKER_URL = "https://github.com/scrapy/scrapyd-client/issues"
+
+
+def deploy(args):  # noqa: ARG001
+    """Deploy a Scrapy project to a Scrapyd instance. For help, invoke scrapyd-deploy."""
+    sys.argv.pop(1)
+    scrapyd_client.deploy.main()
+
+
+def projects(args):
+    """List all projects deployed on a Scrapyd instance."""
+    if _projects := lib.get_projects(args.target, username=args.username, password=args.password):
+        print("\n".join(_projects))
+
+
+def schedule(args):
+    """Schedule the specified spider(s)."""
+    job_args = [(x[0], x[1]) for x in (y.split("=", 1) for y in args.arg)]
+
+    for project in lib.get_projects(args.target, args.project, username=args.username, password=args.password):
+        _spiders = lib.get_spiders(
+            args.target,
+            project,
+            args.spider,
+            username=args.username,
+            password=args.password,
+        )
+        for spider in _spiders:
+            job_id = lib.schedule(
+                args.target,
+                project,
+                spider,
+                job_args,
+                username=args.username,
+                password=args.password,
+            )
+            print(f"{project} / {spider} => {job_id}")
+
+
+def spiders(args):
+    """List all spiders for the given project(s)."""
+    for project in lib.get_projects(args.target, args.project, username=args.username, password=args.password):
+        project_spiders = lib.get_spiders(args.target, project, username=args.username, password=args.password)
+        if not args.verbose:
+            print(f"{project}:")
+            if project_spiders:
+                print(indent("\n".join(project_spiders), INDENT_PREFIX))
+            else:
+                print(f"{INDENT_PREFIX}No spiders.")
+        elif project_spiders:
+            print("\n".join(f"{project} {x}" for x in project_spiders))
 
 
 def parse_cli_args(args):
@@ -52,14 +101,14 @@ def parse_cli_args(args):
         help="Specifies the password to connect to the Scrapyd target.",
     )
 
-    parser = subparsers.add_parser("deploy", description=commands.deploy.__doc__)
-    parser.set_defaults(action=commands.deploy)
+    parser = subparsers.add_parser("deploy", description=deploy.__doc__)
+    parser.set_defaults(action=deploy)
 
-    parser = subparsers.add_parser("projects", description=commands.projects.__doc__)
-    parser.set_defaults(action=commands.projects)
+    parser = subparsers.add_parser("projects", description=projects.__doc__)
+    parser.set_defaults(action=projects)
 
-    parser = subparsers.add_parser("schedule", description=commands.schedule.__doc__)
-    parser.set_defaults(action=commands.schedule)
+    parser = subparsers.add_parser("schedule", description=schedule.__doc__)
+    parser.set_defaults(action=schedule)
     parser.add_argument("-p", "--project", **project_kwargs)
     parser.add_argument(
         "spider",
@@ -73,8 +122,8 @@ def parse_cli_args(args):
         help="Additional argument (key=value), can be specified multiple times.",
     )
 
-    parser = subparsers.add_parser("spiders", description=commands.spiders.__doc__)
-    parser.set_defaults(action=commands.spiders)
+    parser = subparsers.add_parser("spiders", description=spiders.__doc__)
+    parser.set_defaults(action=spiders)
     parser.add_argument("-p", "--project", **project_kwargs)
     parser.add_argument(
         "-v",
@@ -86,7 +135,7 @@ def parse_cli_args(args):
 
     # If 'deploy' is moved to this module, these lines can be removed. (b9ba799)
     parsed_args, _ = mainparser.parse_known_args(args)
-    if getattr(parsed_args, "action", None) is not commands.deploy:
+    if getattr(parsed_args, "action", None) is not deploy:
         parsed_args = mainparser.parse_args(args)
 
     if not hasattr(parsed_args, "action"):