+++ CHANGES.rst
@@ -12,6 +12,9 @@ Added
 Changed
 ^^^^^^^
 
+- **BREAKING CHANGE:** Move exceptions from ``scrapyd_client.utils`` to ``scrapyd_client.exceptions``.
+- **BREAKING CHANGE:** Move ``HEADERS`` from ``scrapyd_client.utils`` to ``scrapyd_client.lib``.
+- **BREAKING CHANGE:** Move ``scrapyd_client.commands`` into ``scrapyd_client.cli``.
 - The ``scrapyd-client schedule`` subcommand accepts multiple ``--arg setting=...`` arguments. (@mxdev88)
 - The ``scrapyd_client.lib.schedule`` and ``scrapyd_client.ScrapyClient.schedule`` methods accept ``args`` as a list, instead of as a dict.
 - The ``scrapyd-deploy --debug`` option prints the subprocess' standard output and standard error, instead of writing to ``stdout`` and ``stderr`` files.
@@ -25,7 +28,8 @@ Fixed
 Removed
 ^^^^^^^
 
-- **BREAKING CHANGE:** Removed ``scrapyd-deploy --list-projects``, in favor of ``scrapyd-client projects``.
+- **BREAKING CHANGE:** Remove ``scrapyd-deploy --list-projects``, in favor of ``scrapyd-client projects``.
+- **BREAKING CHANGE:** Remove ``get_request`` and ``post_request`` from ``scrapyd_client.utils``.
 - Remove ``urllib3`` and ``w3lib`` requirements.
 - Drop support for Python 3.7, 3.8.